+++ ChangeLog.md
@@ -32,6 +32,7 @@ and simplify code. See the Venice *JMH* benchmark *ReflectionBenchmark.java*
 ### Fixed
 
 - fixed function `partition` for overlapping parts (when steps is smaller than n, the size of a part)
+- method execution to throw an exception if a macro is passed where a function is expected